+++ beancount2ledger/__init__.py
@@ -9,6 +9,8 @@ Beancount to Ledger converter
 __license__ = "GPL-2.0-or-later"
 
 from beancount import loader
+from beancount.core import display_context
+from beancount.core.data import filter_txns
 from pkg_resources import DistributionNotFound
 from pkg_resources import get_distribution
 
@@ -21,22 +23,30 @@ except DistributionNotFound:
     __version__ = "undistributed"
 
 
-def convert(entries, output_format="ledger", config={}):
+def convert(entries, output_format="ledger", dcontext=None, config={}):
     """
     Convert beancount entries to ledger output
     """
 
+    if not dcontext:
+        dcontext = display_context.DisplayContext()
+        for entry in filter_txns(entries):
+            for posting in entry.postings:
+                if not posting.units:
+                    continue
+                dcontext.update(posting.units.number, posting.units.currency)
+
     if output_format == "hledger":
-        printer = HLedgerPrinter(config=config)
+        printer = HLedgerPrinter(dcontext=dcontext, config=config)
     else:
-        printer = LedgerPrinter(config=config)
+        printer = LedgerPrinter(dcontext=dcontext, config=config)
     return '\n'.join(printer(entry) for entry in entries)
 
 
-def convert_file(file, output_format="ledger", config={}):
+def convert_file(file, output_format="ledger", dcontext=None, config={}):
     """
     Convert beancount file to ledger output
     """
 
     entries, _, __ = loader.load_file(file)
-    return convert(entries, output_format, config=config)
+    return convert(entries, output_format, dcontext=dcontext, config=config)