+++ tests/hledger_test.py
@@ -395,6 +395,85 @@ class TestHLedgerConversion(test_utils.TestCase):
                 Equity:Opening-Balance
         """, result)
 
+    @loader.load_doc()
+    def test_retain_precision(self, entries, _, ___):
+        """
+            2010-01-01 open Expenses:Nutrition:Food
+            2010-01-01 open Assets:Bank
+            2010-01-01 open Assets:Investments
+
+            2020-01-10 * "Supermarket"
+              Expenses:Nutrition:Food  150.75 THB @ 0.03310116086236 USD
+              Assets:Bank                                      -4.99 USD
+
+            2020-01-10 * "Supermarket"
+              Expenses:Nutrition:Food  150.75 THB @ 0.025207296849 GBP
+              Assets:Bank                                    -3.80 GBP
+
+            2020-02-28 * "Bought GB00BPN5P782"
+              Assets:Investments           1 GB00BPN5P782 {101.689996215 GBP}
+              Assets:Investments                                 -101.69 GBP
+        """
+        result = beancount2ledger.convert(entries, "hledger")
+        self.assertLines(r"""
+            account Expenses:Nutrition:Food
+
+            account Assets:Bank
+
+            account Assets:Investments
+
+            2020-01-10 * Supermarket
+              Expenses:Nutrition:Food                                        150.75 THB @ 0.03310116086236 USD
+              Assets:Bank                                                     -4.99 USD
+
+            2020-01-10 * Supermarket
+              Expenses:Nutrition:Food                                        150.75 THB @ 0.025207296849 GBP
+              Assets:Bank                                                     -3.80 GBP
+
+            2020-02-28 * Bought GB00BPN5P782
+              Assets:Investments                    1 "GB00BPN5P782" @ 101.689996215 GBP
+              Assets:Investments                                             -101.69 GBP
+        """, result)
+
+    @loader.load_doc()
+    def test_rounding(self, entries, _, ___):
+        """
+            2010-01-01 open Expenses:Test
+            2010-01-01 open Assets:Bank
+
+            2020-01-10 * "Test"
+              Expenses:Test                           4.990 USD
+              Assets:Bank                            -4.990 USD
+
+            2020-01-10 * "Test: will fail in ledger due to precision of USD"
+              Expenses:Test            150.75 THB @ 0.03344 USD
+              Assets:Bank                             -5.04 USD
+
+            2020-01-10 * "Test: will fail in ledger due to precision of USD"
+              Expenses:Test               140.1 THB @ 0.021 USD
+              Assets:Bank                             -2.94 USD
+        """
+        result = beancount2ledger.convert(entries, "hledger")
+        self.assertLines(r"""
+            account Expenses:Test
+
+            account Assets:Bank
+
+            2020-01-10 * Test
+              Expenses:Test                                                    4.990 USD
+              Assets:Bank                                                     -4.990 USD
+
+            2020-01-10 * Test: will fail in ledger due to precision of USD
+              Expenses:Test                                                   150.75 THB @ 0.03344 USD
+              Assets:Bank                                                     -5.040 USD
+              Equity:Rounding                                                 -0.001 USD
+
+            2020-01-10 * Test: will fail in ledger due to precision of USD
+              Expenses:Test                                                   140.10 THB @ 0.021 USD
+              Assets:Bank                                                     -2.940 USD
+              Equity:Rounding                                                 -0.002 USD
+        """, result)
+
     def test_example(self):
         """
         Test converted example with hledger